rewrote main menu handling
