+++ portal/src/components/mobile-nav/MobileNav.js
@@ -2,7 +2,7 @@ import React, { useState, useRef, useEffect } from "react";
 import ReactDOM from "react-dom";
 import classnames from "classnames";
 
-import { routingPaths } from "../util/routing";
+import { useMainMenuItems } from "../../useMenuItems";
 import { Xknapp } from "nav-frontend-ikonknapper";
 
 import { GithubLogo } from "../assets/images/svg";
@@ -62,7 +62,7 @@ const MobileNav = ({ ...props }) => {
     // const open =
     //   props.location.pathname.indexOf(route.path) !== -1 && route.routes;
     return (
-      <MobileNavMenuItem open={open} route={route} index={index} key={index}>
+      <MobileNavMenuItem route={route} index={index} key={index}>
         {route.routes && (
           <ul>
             {/* {route.routes
@@ -74,7 +74,7 @@ const MobileNav = ({ ...props }) => {
     );
   };
 
-  const { menuPaths } = routingPaths();
+  const menuItems = useMainMenuItems();
 
   return (
     <div
@@ -95,9 +95,7 @@ const MobileNav = ({ ...props }) => {
           <span className="sr-only">Lukk meny</span>
         </Xknapp>
         <ul className="nav-list">
-          {menuPaths
-            .filter((route) => route.path)
-            .map((route, index) => renderRoute(route, index))}
+          {menuItems.map((route, index) => renderRoute(route, index))}
           <li>
             <a
               href="https://github.com/navikt/nav-frontend-moduler"
