+++ portal/src/pages/accessibility.mdx
@@ -1,5 +1,6 @@
 ---
 title: Tilgjengelighet
+rank: 4
 ---
 
 import NavPage from "../components/NavPage";
