+++ portal/src/pages/components.mdx
@@ -1,5 +1,6 @@
 ---
 title: Komponenter
+rank: 1
 ---
 
 import NavPage from "../components/NavPage";
