+++ portal/src/pages/index.mdx
@@ -1,0 +1,8 @@
+---
+title: Kom i gang
+rank: 0
+---
+
+import GetStartedPage from "../components/get-started/GetStartedPage";
+
+<GetStartedPage />
