+++ portal/src/pages/patterns.mdx
@@ -1,5 +1,6 @@
 ---
 title: Mønster
+rank: 2
 ---
 
 import NavPage from "../components/NavPage";
