+++ portal/src/pages/resources.mdx
@@ -1,5 +1,6 @@
 ---
 title: Ressurser
+rank: 3
 ---
 
 import NavPage from "../components/NavPage";
