+++ portal/src/useMenuItems.tsx
@@ -1,14 +1,14 @@
 import { graphql, useStaticQuery } from "gatsby";
-
-export default (location) => {
-  const sites = useStaticQuery(graphql`
+const useAllMdx = () =>
+  useStaticQuery(graphql`
     query AllMdx {
-      allMdx {
+      allMdx(sort: { fields: frontmatter___rank }) {
         edges {
           node {
             slug
             frontmatter {
               title
+              rank
             }
           }
         }
@@ -16,8 +16,21 @@ export default (location) => {
     }
   `);
 
-  return sites.allMdx.edges
-    .filter(
+export const useMainMenuItems = () =>
+  useAllMdx()
+    .allMdx.edges.filter(
+      (edge) =>
+        edge.node.frontmatter.rank !== null &&
+        edge.node.slug.split("/").length === 1
+    )
+    .map((edge) => ({
+      link: `/${edge.node.slug}`,
+      title: edge.node.frontmatter.title,
+    }));
+
+export default (location) =>
+  useAllMdx()
+    .allMdx.edges.filter(
       (edge) =>
         edge.node.slug.split("/").length === 2 &&
         edge.node.slug.split("/")[0].startsWith(location.pathname.split("/")[1])
@@ -26,4 +39,3 @@ export default (location) => {
       link: `/${edge.node.slug}`,
       title: edge.node.frontmatter.title,
     }));
-};
